Update site to use chirpy `v6.5.0`
Update the Gemfile to use the latest version of the jekyll-theme-chirpy gem. Also, add a new Gemfile.lock file with the updated revision of the jekyll-theme-chirpy gem.

The Gemfile now includes the line `gem "jekyll-theme-chirpy", "~> 6.5.0", git: "https://github.com/powerium/jekyll-theme-powerium.git"`, which specifies the updated version of the gem.

A new Gemfile.lock file has been created, which includes the revision 28e073f49bf4f6e7465a3e5e9c5947c68ff64e2d of the jekyll-theme-chirpy gem.

The changes ensure that the jekyll-theme-chirpy gem is up to date and the application is using the specified version of the gem.

Update Gemfile and Gemfile.lock with the latest version of jekyll-theme-chirpy.